# swagger.json
{"openapi":"3.0.0","paths":{"/api/auth/login":{"post":{"operationId":"AuthController_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthAdminInputDto"}}}},"responses":{"201":{"description":""}},"tags":["Auth"]}},"/api/device":{"get":{"operationId":"DeviceController_getDevices","parameters":[{"name":"page","required":false,"in":"query","schema":{"example":1,"type":"number"}},{"name":"limit","required":false,"in":"query","schema":{"example":10,"type":"number"}}],"responses":{"200":{"description":""}},"security":[{"authorization":[]}],"tags":["Device"]}},"/api/device/register":{"post":{"operationId":"DeviceController_register","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterDeviceInputDto"}}}},"responses":{"201":{"description":""}},"tags":["Device"]}},"/api/news":{"get":{"operationId":"NewsController_getAll","parameters":[{"name":"page","required":false,"in":"query","schema":{"example":1,"type":"number"}},{"name":"limit","required":false,"in":"query","schema":{"example":10,"type":"number"}},{"name":"type","required":false,"in":"query","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"security":[{"authorization":[]}],"tags":["News"]}},"/api/news/check-news":{"get":{"operationId":"NewsController_checkNews","parameters":[{"name":"Device-Id","required":true,"in":"header","description":"Unique ID of the device","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["News"]}},"/api/news/banner":{"post":{"operationId":"NewsController_createBanner","parameters":[],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/CreateBannerInputDto"}}}},"responses":{"201":{"description":""}},"security":[{"authorization":[]}],"tags":["News"]}},"/api/news/alert":{"post":{"operationId":"NewsController_createAlert","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateAlertInputDto"}}}},"responses":{"201":{"description":""}},"security":[{"authorization":[]}],"tags":["News"]}},"/api/news/alert-update":{"post":{"operationId":"NewsController_createUpdateAlert","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUpdateAlertInputDto"}}}},"responses":{"201":{"description":""}},"security":[{"authorization":[]}],"tags":["News"]}},"/api/news/change-status/{id}":{"post":{"operationId":"NewsController_changeStatus","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"201":{"description":""}},"security":[{"authorization":[]}],"tags":["News"]}}},"info":{"title":"Podium News Manager","description":"The Podium News Manager API description","version":"0.0.1","contact":{"name":"Muhammadali","url":"http://yuldoshev.uz","email":"mukhammadaliweb@gmail.com"}},"tags":[],"servers":[],"components":{"securitySchemes":{"authorization":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"AuthAdminInputDto":{"type":"object","properties":{"username":{"type":"string","example":"admin"},"password":{"type":"string","example":"12345678"}},"required":["username","password"]},"RegisterDeviceInputDto":{"type":"object","properties":{"app_version":{"type":"string","example":"1.0.0"},"token":{"type":"string","example":"12345678"},"locate":{"type":"string","example":"uz"},"timezone":{"type":"number","example":5},"os":{"type":"number","example":1}},"required":["app_version","token","locate","timezone","os"]},"CreateBannerInputDto":{"type":"object","properties":{"image":{"type":"string","format":"binary"}},"required":["image"]},"CreateAlertInputDto":{"type":"object","properties":{"title":{"type":"string","example":"Promotion!!!"},"body":{"type":"string","example":"We have a promotion going on"}},"required":["title","body"]},"CreateUpdateAlertInputDto":{"type":"object","properties":{"title":{"type":"string","example":"An update has been released"},"body":{"type":"string","example":"Download"},"url":{"example":["https://example.com","https://example.com"],"type":"array","items":{"type":"string"}}},"required":["title","body","url"]}}}}